backend/api-server/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.0.0
python-multipart>=0.0.6
pillow>=10.0.0
pillow-heif>=0.13.0
google-genai>=1.0.0
